+++ scripts/mod_00_x=Omega_l_v=ams.py
@@ -16,7 +16,7 @@ params = {
         'show_progress': True,
         'X': {
             'var': 'Omegas',
-            'idx': 1,
+            'idx': 0,
             'min': 1.0,
             'max': 2.0,
             'dim': 51
@@ -24,12 +24,10 @@ params = {
     },
     'solver': {
         'cache': True,
-        'cache_dir': 'H:/Workspace/VSCode/Python/entan_oem_mod/data/mod_00/0.0_1000.0_10001',
-        'method': 'ode',
-        'measure_type': 'qcm',
-        'qcm_type': 'entan',
-        'idx_mode_i': 1,
-        'idx_mode_j': 2,
+        'cache_dir': 'H:/Workspace/data/mod_00/0.0_1000.0_10001',
+        'method': 'zvode',
+        'measure_type': 'entan_ln',
+        'idx_e': (1, 2),
         'range_min': 9371,
         'range_max': 10001,
         't_min': 0,
@@ -46,12 +44,12 @@ params = {
         'n_ths': [0, 0],
         'Omegas': [2.0, 2.0],
         'omegas': [1.0, 1.0],
-        't_mod': 'cos',
+        't_mods': ['cos', 'cos'],
         't_pos': 'bottom'
     },
     'plotter': {
         'type': 'line',
-        'x_label': '$\\Omega_{v} / \\omega_{0}$',
+        'x_label': '$\\Omega_{l} / \\omega_{0}$',
         'x_bound': 'both',
         'x_ticks': [1.0, 1.2, 1.4, 1.6, 1.8, 2.0],
         'v_label': '$E_{N}$',
@@ -61,7 +59,6 @@ params = {
 }
 
 # get average entanglement
-looper = wrap_looper(Mod00, params, 'measure_average', 'XLooper', 'H:/Workspace/VSCode/Python/entan_oem_mod/data/mod_00/E_N_g_0_0.005_g_1_0.005_Omega_l_1.16', True)
-
+looper = wrap_looper(system=Mod00, params=params, func='ams', looper='x_looper', file_path='data/mod_00/entan_ln', plot=True)
 # calculate thresholds
-print(looper.get_thresholds())
+print(looper.get_thresholds(thres_mode='minmax'))