+++ scripts/mod_00_x=Omega_v_v=ams.py
@@ -13,14 +13,8 @@ from systems.Mod00 import Mod00
 # all parameters
 params = {
     'looper': {
+        'show_progress': True,
         'X': {
-            'var': 'Omegas',
-            'idx': 0,
-            'min': 1.0,
-            'max': 2.0,
-            'dim': 51
-        },
-        'Y': {
             'var': 'Omegas',
             'idx': 1,
             'min': 1.0,
@@ -30,12 +24,10 @@ params = {
     },
     'solver': {
         'cache': True,
-        'cache_dir': 'H:/Workspace/VSCode/Python/entan_oem_mod/data/mod_00/0.0_1000.0_10001',
-        'method': 'ode',
-        'measure_type': 'qcm',
-        'qcm_type': 'entan',
-        'idx_mode_i': 1,
-        'idx_mode_j': 2,
+        'cache_dir': 'H:/Workspace/data/mod_00/0.0_1000.0_10001',
+        'method': 'zvode',
+        'measure_type': 'entan_ln',
+        'idx_e': (1, 2),
         'range_min': 9371,
         'range_max': 10001,
         't_min': 0,
@@ -52,28 +44,21 @@ params = {
         'n_ths': [0, 0],
         'Omegas': [2.0, 2.0],
         'omegas': [1.0, 1.0],
-        't_mod': 'cos',
+        't_mods': ['cos', 'cos'],
         't_pos': 'bottom'
     },
     'plotter': {
-        'type': 'contourf',
-        'palette': 'RdBu',
-        'bins': 11,
-        'x_label': '$\\Omega_{l} / \\omega_{0}$',
+        'type': 'line',
+        'x_label': '$\\Omega_{v} / \\omega_{0}$',
         'x_bound': 'both',
         'x_ticks': [1.0, 1.2, 1.4, 1.6, 1.8, 2.0],
-        'y_label': '$\\Omega_{v} / \\omega_{1}$',
-        'y_bound': 'both',
-        'y_ticks': [1.0, 1.2, 1.4, 1.6, 1.8, 2.0],
-        'show_cbar': True,
-        'cbar_position': 'top',
-        'cbar_title': '$E_{N}$',
-        'cbar_ticks': [0.1, 0.2, 0.3, 0.4]
+        'v_label': '$E_{N}$',
+        'v_bound': 'both',
+        'v_ticks': [0.0, 0.2, 0.4, 0.6]
     }
 }
 
 # get average entanglement
-looper = wrap_looper(Mod00, params, 'measure_average', 'XYLooper', 'H:/Workspace/VSCode/Python/entan_oem_mod/data/mod_00/E_N_g_0_0.005_g_1_0.005', True)
-
+looper = wrap_looper(system=Mod00, params=params, func='ams', looper='x_looper', file_path='data/mod_00/entan_ln', plot=True)
 # calculate thresholds
-print(looper.get_thresholds())
+print(looper.get_thresholds(thres_mode='minmax'))