+++ systems/Mod00.py
@@ -5,7 +5,7 @@
 
 __authors__ = ['Sampreet Kalita']
 __created__ = '2021-05-18'
-__updated__ = '2021-06-26'
+__updated__ = '2021-08-04'
 
 # dependencies
 import numpy as np
@@ -42,7 +42,7 @@ class Mod00(SODMSystem):
             'n_ths': params.get('n_ths', [0, 0]),
             'Omegas': params.get('Omegas', [2.0, 2.0]),
             'omegas': params.get('omegas', [1.0, 1.0]),
-            't_mod': params.get('t_mod', 'cos'),
+            't_mods': params.get('t_mod', ['cos', 'cos']),
             't_pos': params.get('t_pos', 'top')
         }
         # drift matrix
@@ -54,7 +54,7 @@ class Mod00(SODMSystem):
         Parameters
         ----------
         modes : list
-            Values of the optical and mechancial modes.
+            Values of the modes.
         params : list
             Constant parameters.
         t : float
@@ -72,7 +72,7 @@ class Mod00(SODMSystem):
         gs      = [params[7], params[8]]
         kappa   = params[9]
         omegas  = [params[12], params[13]]
-        t_pos   = params[15]
+        t_pos   = params[16]
         alpha   = modes[0]
         betas   = [modes[1], modes[2]]
 
@@ -136,7 +136,7 @@ class Mod00(SODMSystem):
             Next element contains the optical decay rate.
             Next 2 elements contain the laser and voltage modulation frequencies.
             Next 2 elements contain the mechanical and LC frequencies.
-            Next element contains the type of modulation function.
+            Next 2 elements contain the type of modulation function.
             Next element contains the position of the mechanical element.
         """
 
@@ -150,7 +150,7 @@ class Mod00(SODMSystem):
         n_ths   = self.params['n_ths']
         Omegas  = self.params['Omegas']
         omegas  = self.params['omegas']
-        t_mod   = self.params['t_mod']
+        t_mods  = self.params['t_mods']
         t_pos   = self.params['t_pos']
  
         # initial mode values as 1D list
@@ -158,12 +158,12 @@ class Mod00(SODMSystem):
 
         # initial quadrature correlations
         corrs_0 = np.zeros([6, 6], dtype=np.float_)
-        corrs_0[0][0] = 1/2 
-        corrs_0[1][1] = 1/2
-        corrs_0[2][2] = (n_ths[0] + 1/2)
-        corrs_0[3][3] = (n_ths[0] + 1/2)
-        corrs_0[4][4] = (n_ths[1] + 1/2)
-        corrs_0[5][5] = (n_ths[1] + 1/2)
+        corrs_0[0][0] = 0.5 
+        corrs_0[1][1] = 0.5
+        corrs_0[2][2] = (n_ths[0] + 0.5)
+        corrs_0[3][3] = (n_ths[0] + 0.5)
+        corrs_0[4][4] = (n_ths[1] + 0.5)
+        corrs_0[5][5] = (n_ths[1] + 0.5)
 
         # convert to 1D list and concatenate all variables
         iv = modes_0 + [np.complex(element) for element in corrs_0.flatten()]
@@ -189,7 +189,8 @@ class Mod00(SODMSystem):
             [kappa] + \
             Omegas + \
             omegas + \
-            [t_mod, t_pos]
+            t_mods + \
+            [t_pos]
 
         # all constants
         c = D.flatten().tolist() + params
@@ -223,27 +224,28 @@ class Mod00(SODMSystem):
         kappa   = params[9]
         Omegas  = [params[10], params[11]]
         omegas  = [params[12], params[13]]
-        t_mod   = params[14]
-        t_pos   = params[15]
+        t_mods  = [params[14], params[15]]
+        t_pos   = params[16]
         alpha   = modes[0]
         betas   = [modes[1], modes[2]]
 
         # effective values
         Delta = Delta_0 - 2 * gs[0] * np.real(betas[0])
         g_1 = - gs[1] if t_pos == 'bottom' else gs[1] 
-        func_mod = lambda Theta : {
-            'exp': np.exp(1j * Theta),
-            'cos': np.cos(Theta),
-            'sin': np.sin(Theta)
-        }.get(t_mod, np.cos(Theta))
+        # modulation function selector
+        get_mod = lambda idx, omega : {
+            'cos': np.cos(omega * t),
+            'exp': np.exp(1j * omega * t),
+            'sin': np.sin(omega * t)
+        }.get(t_mods[idx], np.cos(omega * t))
 
         # calculate rates
         # optical mode
-        dalpha_dt = - (kappa + 1j * Delta) * alpha + A_ls[0] + A_ls[1] * func_mod(Omegas[0] * t)
+        dalpha_dt = - (kappa + 1j * Delta) * alpha + A_ls[0] + A_ls[1] * get_mod(0, Omegas[0])
         # mechanical mode
         dbeta_0_dt = 1j * gs[0] * np.conjugate(alpha) * alpha - (gammas[0] + 1j * omegas[0]) * betas[0] + 4j * g_1 * np.real(betas[1])**2
         # circuit mode
-        dbeta_1_dt = 8j * g_1 * np.real(betas[0]) * np.real(betas[1]) - (gammas[1] + 1j * omegas[1]) * betas[1] + 1j * (A_vs[0] + A_vs[1] * func_mod(Omegas[1] * t))
+        dbeta_1_dt = 8j * g_1 * np.real(betas[0]) * np.real(betas[1]) - (gammas[1] + 1j * omegas[1]) * betas[1] + 1j * (A_vs[0] + A_vs[1] * get_mod(1, Omegas[1]))
 
         # arrange rates
         mode_rates = [dalpha_dt, dbeta_0_dt, dbeta_1_dt]